changed the sample to use file share databus with learning transport
