--- samples/databus/blob-storage-databus-custom-serializer-converter/ABSDataBus_5/AzureBlobStorageDataBus.sln
@@ -1,27 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
-MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sender", "Sender\Sender.csproj", "{58F9345A-2622-4EC5-9B74-1DF74E639FB3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Receiver", "Receiver\Receiver.csproj", "{44A0ED15-DEC7-48BA-810E-A08FFD0DBA46}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "Shared\Shared.csproj", "{5B1F434B-952F-43B8-A10D-BBB622403971}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{58F9345A-2622-4EC5-9B74-1DF74E639FB3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{58F9345A-2622-4EC5-9B74-1DF74E639FB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{44A0ED15-DEC7-48BA-810E-A08FFD0DBA46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{44A0ED15-DEC7-48BA-810E-A08FFD0DBA46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5B1F434B-952F-43B8-A10D-BBB622403971}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5B1F434B-952F-43B8-A10D-BBB622403971}.Debug|Any CPU.Build.0 = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
