--- samples/databus/blob-storage-databus-custom-serializer-converter/ABSDataBus_5/Shared/MessageWithLargePayload.cs
@@ -1,14 +1,0 @@
-﻿using NServiceBus;
-
-#region MessageWithLargePayload
-
-[TimeToBeReceived("00:03:00")]
-public class MessageWithLargePayload :
-    ICommand
-{
-    public DataBusProperty<byte[]> LargePayload { get; set; }
-    public string Description { get; set; }
-}
-
-#endregion
-
