+++ samples/databus/file-share-databus-custom-serializer-converter/Core_8/Receiver/Receiver.csproj
@@ -5,8 +5,6 @@
     <LangVersion>10.0</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.*" />
-    <PackageReference Include="NServiceBus.DataBus.AzureBlobStorage" Version="5.*" />
     <PackageReference Include="NServiceBus" Version="8.*" />
     <ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
