+++ samples/databus/file-share-databus-custom-serializer-converter/Core_8/Sender/Program.cs
@@ -1,0 +1,67 @@
+using NServiceBus;
+using Shared;
+using System;
+using System.Text.Json;
+using System.Threading.Tasks;
+
+class Program
+{
+    static async Task Main()
+    {
+        Console.Title = "Samples.DataBus.Sender";
+        var endpointConfiguration = new EndpointConfiguration("Samples.DataBus.Sender");
+
+        #region ConfigureDataBus
+
+        var dataBus = endpointConfiguration.UseDataBus<FileShareDataBus, SystemJsonDataBusSerializer>();
+        dataBus.BasePath(@"..\..\..\..\storage");
+
+        #endregion
+
+        endpointConfiguration.UsePersistence<LearningPersistence>();
+        #region CustomJsonSerializerOptions
+        var jsonSerializerOptions = new JsonSerializerOptions();
+        jsonSerializerOptions.Converters.Add(new DatabusPropertyConverterFactory());
+        endpointConfiguration.UseSerialization<SystemJsonSerializer>().Options(jsonSerializerOptions);
+        #endregion
+
+        endpointConfiguration.UseTransport(new LearningTransport());
+        var endpointInstance = await Endpoint.Start(endpointConfiguration)
+            .ConfigureAwait(false);
+        Console.WriteLine("Press 'D' to send a databus large message");
+        Console.WriteLine("Press any other key to exit");
+
+        while (true)
+        {
+            var key = Console.ReadKey();
+            Console.WriteLine();
+
+            if (key.Key == ConsoleKey.D)
+            {
+                await SendMessageLargePayload(endpointInstance)
+                    .ConfigureAwait(false);
+                continue;
+            }
+            break;
+        }
+        await endpointInstance.Stop()
+            .ConfigureAwait(false);
+    }
+
+    static async Task SendMessageLargePayload(IEndpointInstance endpointInstance)
+    {
+        #region SendMessageLargePayload
+
+        var message = new MessageWithLargePayload
+        {
+            SomeProperty = "This message contains a large blob that will be sent on the data bus",
+            LargeBlob = new DataBusProperty<byte[]>(new byte[1024*1024*5]) //5MB
+        };
+        await endpointInstance.Send("Samples.DataBus.Receiver", message)
+            .ConfigureAwait(false);
+
+        #endregion
+
+        Console.WriteLine(@"Message sent, the payload is stored in: ..\..\..\storage");
+    }
+}
