+++ samples/databus/file-share-databus-custom-serializer-converter/Core_8/Sender/Sender.csproj
@@ -5,9 +5,7 @@
     <LangVersion>10.0</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.*" />
     <PackageReference Include="NServiceBus" Version="8.*" />
-    <PackageReference Include="NServiceBus.DataBus.AzureBlobStorage" Version="5.*" />
     <ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
 </Project>
