+++ samples/databus/file-share-databus-custom-serializer-converter/Core_8/Shared/MessageWithLargePayload.cs
@@ -1,0 +1,14 @@
+﻿using NServiceBus;
+
+#region MessageWithLargePayload
+
+//the data bus is allowed to clean up transmitted properties older than the TTBR
+[TimeToBeReceived("00:01:00")]
+public class MessageWithLargePayload :
+    ICommand
+{
+    public string SomeProperty { get; set; }
+    public DataBusProperty<byte[]> LargeBlob { get; set; }
+}
+
+#endregion
