+++ samples/databus/file-share-databus-custom-serializer-converter/Core_8/Shared/Shared.csproj
